Removed compilation warning
git-svn-id: http://svn.opennebula.org/one/trunk@392 3034c82b-c49b-4eb3-8279-a7acafdc01c0

+++ src/vm/VirtualMachinePool.cc
@@ -86,7 +86,7 @@ int VirtualMachinePool::allocate (
     // ------------------------------------------------------------------------
     generate_context(*oid,attrs);
 
-    for (int i = 0; i < attrs.size() ; i++)
+    for (int i = 0; i < (int) attrs.size() ; i++)
     {
         if (attrs[i] != 0)
         {
